Refactor: http-testserver 2.0
   To support new microservice architecture refactoring which is based
   on junit 5 for unit and integration testing, did the following
   refactoring:

   * Refatcoring to two modules: core and junit4
   * Core module is junit independent, only includes base http test server
     implementation
   * Junit4 module implements rules based on core module
   * Future will have junit5 module to support junit 5 extensions
