+++ .gitignore
@@ -1,0 +1,22 @@
+# build
+*.tar.gz
+dependency-reduced-pom.xml
+scratch
+target
+
+# eclipse
+.classpath
+.project
+.settings
+
+# idea
+.idea
+*.iml
+*.iws
+*.ipr
+
+# mac
+*.DS_Store
+
+
+*.dat
