+++ core/src/main/java/org/commonjava/test/http/common/CommonMethod.java
@@ -1,0 +1,28 @@
+/**
+ * Copyright (C) 2011-2021 Red Hat, Inc. (http://github.com/Commonjava/http-testserver)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.commonjava.test.http.common;
+
+/**
+ * Convenience enum. Some of the more common HTTP methods in use.
+ * 
+ * @author jdcasey
+ */
+public enum CommonMethod
+{
+
+    GET, PUT, POST, DELETE, HEAD, OPTIONS
+
+}
