--- src/main/java/org/commonjava/test/http/stream/StreamResolver.java
@@ -1,28 +1,0 @@
-/**
- * Copyright (C) 2015 Red Hat, Inc. (jdcasey@commonjava.org)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.commonjava.test.http.stream;
-
-import java.io.IOException;
-import java.io.InputStream;
-
-/**
- * Created by jdcasey on 8/17/15.
- */
-public interface StreamResolver
-{
-    InputStream get( String path )
-            throws IOException;
-}
